feat: ゲームの基本機能を実装（プレイヤー、敵、コイン、オーディオ、タッチ操作、ステージシステム）
