+++ index.html
@@ -42,6 +42,6 @@
 </head>
 <body>
     <canvas id="gameCanvas"></canvas>
-    <script type="module" src="dist/game.js"></script>
+    <script type="module" src="dist/main.js"></script>
 </body>
 </html>
