+++ src/main.ts
@@ -1,0 +1,5 @@
+import { initGame } from './game.js';
+
+window.onload = () => {
+    initGame();
+};
